+++ app/afterlogin/account/page.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react"
 import Image from "next/image"
 import { Card } from "../../../components/aftelogin"
 import Logo from "/public/Logo.png"
-import { useEthersContext } from "../../../components/aftelogin/contexts/EthersContext"
+import { useEthersContext } from "../../../context/EthersContext"
 
 const Account = () => {
   const { signer } = useEthersContext()
